Agregando clases repository y servicios para las operaciones de la base de datos

+++ MELI/src/main/java/com/freecupon/MELI/domain/dto/ItemsExchangedDto.java
@@ -9,6 +9,6 @@ import lombok.Setter;
 public class ItemsExchangedDto {
 
     private String item_id;
-    private Integer exchanged;
+    private Long exchanged;
 
 }
